+++ ql/src/java/org/apache/hadoop/hive/ql/parse/UpdateDeleteSemanticAnalyzer.java
@@ -294,7 +294,7 @@ public class UpdateDeleteSemanticAnalyzer extends RewriteSemanticAnalyzer {
   }
 
   @Override
-  protected boolean enableColumnStatsCollecting() {
+  public boolean enableColumnStatsCollecting() {
     return false;
   }
 }
